tests: separate callbacks from app html definitions
